fix typo and a cast
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3951 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Frameworks/Addresses/ADMultiValue.m
@@ -180,7 +180,7 @@ static ADPropertyType _propTypeFromDict(NSDictionary *dict)
   while((dict = [e nextObject]))
     max = MAX(max, [[dict objectForKey: @"ID"] integerValue]);
   
-  return [NSString stringWithFormat: @"%d", max+1];
+  return [NSString stringWithFormat: @"%ld", (long int)(max+1)];
 }
 
 - (NSString*) addValue: (id) value
